Move scipp_dtype to the proper place.
