+++ python/dtype.h
@@ -1,0 +1,5 @@
+namespace pybind11 {
+class dtype;
+}
+
+scipp::core::DType scipp_dtype(const pybind11::dtype &type);
